requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai>=0.2.9
langchain-community
pandas
numpy
pydantic
python-dotenv
psycopg2-binary
openai>=1.40.0
httpx>=0.27.0